+++ docs/diagrams/feature-flag-system.mmd
@@ -1,28 +1,55 @@
 flowchart TD
-    A[User Request] --> B{Check Feature Flags}
-    B -->|ENABLE_APP = true| C[Show App Route]
-    B -->|ENABLE_APP = false| D[Hide App Route]
+    %% Feature Flag System Diagram
+    %% Shows how feature flags control application behavior
     
-    C --> E[Render Navigation]
+    %% Define styles
+    classDef userFlow fill:#f9f,stroke:#333,stroke-width:2px
+    classDef featureFlag fill:#bbf,stroke:#333,stroke-width:1px
+    classDef component fill:#bfb,stroke:#333,stroke-width:1px
+    classDef decision fill:#ffd,stroke:#333,stroke-width:1px,shape:diamond
+    
+    %% User flow
+    A[User Request]:::userFlow --> B{Check Feature Flags}:::decision
+    B -->|ENABLE_APP = true| C[Show App Route]:::userFlow
+    B -->|ENABLE_APP = false| D[Hide App Route]:::userFlow
+    
+    C --> E[Render Navigation]:::component
     D --> E
     
-    subgraph Feature Flag System
-    F[Statsig Client] --> G[useGateValue Hook]
-    G --> B
+    %% Feature Flag System components
+    subgraph "Feature Flag System" 
+        direction TB
+        F[Statsig Client]:::featureFlag --> G[useGateValue Hook]:::featureFlag
+        G --> B
+        
+        %% Add feature flag examples
+        H[Feature Flags]:::featureFlag
+        H --> |Controls| I[ENABLE_APP]:::featureFlag
+        H --> |Controls| J[ENABLE_SEARCH]:::featureFlag
+        H --> |Controls| K[ENABLE_PROFILE]:::featureFlag
     end
     
-    subgraph Navigation Components
-    E --> H[DesktopNav]
-    E --> I[MobileNav]
-    E --> J[Footer]
+    %% Navigation components
+    subgraph "Navigation Components"
+        direction LR
+        E --> H1[DesktopNav]:::component
+        E --> I1[MobileNav]:::component
+        E --> J1[Footer]:::component
     end
     
-    subgraph useNavItems Hook
-    K[filterNavItems] --> L{Check Feature Flags}
-    L -->|Enabled| M[Include Route]
-    L -->|Disabled| N[Exclude Route]
+    %% Nav items filtering
+    subgraph "useNavItems Hook"
+        direction TB
+        K1[filterNavItems]:::component --> L{Check Feature Flags}:::decision
+        L -->|Enabled| M[Include Route]:::component
+        L -->|Disabled| N[Exclude Route]:::component
     end
     
-    H --> K
-    I --> K
-    J --> K 
+    %% Connect components
+    H1 --> K1
+    I1 --> K1
+    J1 --> K1
+    
+    %% Add notes
+    note["Note: Feature flags can be updated remotely<br>without requiring a new deployment"]
+    style note fill:#ffffcc,stroke:#999,stroke-width:1px,stroke-dasharray: 5 5
