+++ docs/diagrams/places-api-flow.mmd
@@ -1,24 +1,44 @@
 sequenceDiagram
-    participant Client
-    participant API as Next.js API Route
-    participant Cache as Redis Cache
-    participant Google as Google Places API
+    %% Places API Flow Diagram
+    %% Shows the sequence of operations when fetching places
     
-    Client->>API: GET /api/places/nearby/places-new
+    %% Define participants with styling
+    participant Client as 🌐 Client
+    participant API as 🔄 Next.js API Route
+    participant Cache as 💾 Redis Cache
+    participant Google as 🔍 Google Places API
     
-    Note over API: Process request parameters
+    %% Add styling
+    autonumber
     
-    API->>Cache: Check cache for results
+    %% Request flow
+    Client->>+API: GET /api/places/nearby/places-new
+    Note over API: Process request parameters<br/>(location, radius, keywords)
     
+    %% Cache check
+    API->>+Cache: Check cache for results
+    
+    %% Conditional flow based on cache
     alt Cache hit
-        Cache-->>API: Return cached results
+        Cache-->>-API: Return cached results
+        API-->>-Client: Return processed places (from cache)
+        Note right of Client: Fast response (< 100ms)
     else Cache miss
-        API->>Google: Fetch from Google Places API
-        Google-->>API: Return places data
+        Cache-->>-API: No cached data
+        API->>+Google: Fetch from Google Places API
+        Note over Google: Search for places<br/>based on parameters
+        Google-->>-API: Return places data
         
-        Note over API: Process places with<br/>findMatchingKeyword<br/>createSimplifiedPlace
+        rect rgb(240, 248, 255)
+            Note over API: Data Processing
+            API->>API: Process places with utility functions
+            Note over API: findMatchingKeyword()<br/>createSimplifiedPlace()
+        end
         
-        API->>Cache: Store results in cache
+        API->>Cache: Store results in cache (TTL: 7 days)
+        API-->>-Client: Return processed places (from API)
+        Note right of Client: Slower response (300-800ms)
     end
     
-    API-->>Client: Return processed places 
+    %% Add a note about caching
+    Note over Cache: Cache key format:<br/>places:{location}:{radius}:{keywords}
