+++ src/components/nav/desktop-nav/Desktop-Nav.tsx
@@ -47,17 +47,19 @@ export function DesktopNav({ navItems }: DesktopNavProps) {
                 data-testid='nav-child'
                 className='absolute left-0 mt-2 w-48 bg-[#34409b] dark:bg-[#34409b] shadow-lg rounded-b-lg overflow-hidden z-10 opacity-0 group-hover:opacity-100 group-hover:translate-y-1 transition-opacity duration-200 delay-150'
               >
-                {navItem.children.filter(shouldShowNavItem).map((child, index) => (
-                  <Link
-                    key={`${child.label}-${index}`}
-                    href={child.href}
-                    target={child.target ? '_blank' : undefined}
-                    rel={child.target ? 'noopener noreferrer' : undefined}
-                    className='block px-4 py-2 text-white/75 hover:text-white hover:bg-[#2a3480] dark:hover:bg-[#2a3480] text-sm transition-colors'
-                  >
-                    {child.label}
-                  </Link>
-                ))}
+                {navItem.children
+                  .filter(shouldShowNavItem)
+                  .map((child, index) => (
+                    <Link
+                      key={`${child.label}-${index}`}
+                      href={child.href}
+                      target={child.target ? '_blank' : undefined}
+                      rel={child.target ? 'noopener noreferrer' : undefined}
+                      className='block px-4 py-2 text-white/75 hover:text-white hover:bg-[#2a3480] dark:hover:bg-[#2a3480] text-sm transition-colors'
+                    >
+                      {child.label}
+                    </Link>
+                  ))}
               </div>
             )}
           </div>
