+++ src/components/nav/mobile-nav/Mobile-Nav.test.tsx
@@ -142,15 +142,16 @@ describe('MobileNav', () => {
     // Setup default mock implementation
     (navHooks.useNavItems as ReturnType<typeof vi.fn>).mockReturnValue({
       shouldShowNavItem: mockShouldShowNavItem,
-      filterNavItems: (items: NavItem[]) => items.filter(mockShouldShowNavItem).map(item => {
-        if (item.children) {
-          return {
-            ...item,
-            children: item.children.filter(mockShouldShowNavItem)
-          };
-        }
-        return item;
-      }),
+      filterNavItems: (items: NavItem[]) =>
+        items.filter(mockShouldShowNavItem).map((item) => {
+          if (item.children) {
+            return {
+              ...item,
+              children: item.children.filter(mockShouldShowNavItem),
+            };
+          }
+          return item;
+        }),
     });
   });
 
@@ -216,15 +217,18 @@ describe('MobileNav', () => {
         if (item.hidden) return false;
         return true; // All feature flags enabled
       },
-      filterNavItems: (items: NavItem[]) => items.filter(item => !item.hidden).map(item => {
-        if (item.children) {
-          return {
-            ...item,
-            children: item.children.filter(child => !child.hidden)
-          };
-        }
-        return item;
-      }),
+      filterNavItems: (items: NavItem[]) =>
+        items
+          .filter((item) => !item.hidden)
+          .map((item) => {
+            if (item.children) {
+              return {
+                ...item,
+                children: item.children.filter((child) => !child.hidden),
+              };
+            }
+            return item;
+          }),
     });
 
     render(<MobileNav navItems={mockNavItems} />);
